Mocha is not thread-safe, and unit tests should never invoke kubectl
